Closer to somewhat functional. Can't figure out how to append arguments yet.

+++ mpm.go
@@ -7,16 +7,21 @@ import (
 	"io"
 	"net/http"
 	"os"
+	"os/exec"
 	"path/filepath"
 	"runtime"
 	"strings"
+
+	"github.com/fatih/color"
 )
 
 func main() {
 	var operatingSystem string
 	var defaultTMP string
+	var mpmDownloadPath string
 	var mpmURL string
 	var mpmDownloadNeeded bool
+	var mpmExtractNeeded bool
 	var release string
 	var defaultPath string
 
@@ -50,12 +55,13 @@ func main() {
 	fmt.Println("Operating System:", operatingSystem)
 	fmt.Println("MPM URL:", mpmURL)
 	mpmDownloadNeeded = true
+	mpmExtractNeeded = true
+	red := color.New(color.FgRed).SprintFunc()
 
 	// Figure out where you want actual MPM to go.
 	for {
 		fmt.Print("Enter the path to the directory where you would like MPM to download to. " +
-			"Press Enter to use \"" + defaultTMP + "\": ")
-		var mpmDownloadPath string
+			"Press Enter to use \"" + defaultTMP + "\"\n> ")
 		fmt.Scanln(&mpmDownloadPath)
 
 		if mpmDownloadPath == "" {
@@ -63,13 +69,18 @@ func main() {
 		} else {
 			_, err := os.Stat(mpmDownloadPath)
 			if os.IsNotExist(err) {
-				fmt.Printf("The directory \"%s\" does not exist. Do you want to create it? (y/n): ", mpmDownloadPath)
+				fmt.Printf("The directory \"%s\" does not exist. Do you want to create it? (y/n)\n> ", mpmDownloadPath)
 				var createDir string
 				fmt.Scanln(&createDir)
+
+				if createDir == "exit" || createDir == "Exit" || createDir == "quit" || createDir == "Quit" {
+					os.Exit(0)
+				}
+
 				if createDir == "y" || createDir == "Y" {
 					err := os.MkdirAll(mpmDownloadPath, 0755)
 					if err != nil {
-						fmt.Println("Failed to create the directory:", err, "Please select a different directory.")
+						fmt.Println(red("Failed to create the directory:", err, "Please select a different directory."))
 						continue
 					}
 					fmt.Println("Directory created successfully.")
@@ -78,7 +89,7 @@ func main() {
 					continue
 				}
 			} else if err != nil {
-				fmt.Println("Error checking the directory:", err, "Please select a different directory.")
+				fmt.Println(red("Error checking the directory:", err, "Please select a different directory."))
 				continue
 			}
 		}
@@ -86,24 +97,36 @@ func main() {
 		// Check if MPM already exists in the selected directory.
 		fileName := mpmDownloadPath + "/mpm"
 		_, err := os.Stat(fileName)
-		if err == nil {
-			fmt.Printf("MPM already exists in this directory. Would you like to overwrite it? " +
-				"This will also overwrite the directory 'mpm-contents' if it already exists. (y/n): ")
-			var overwriteMPM string
-			fmt.Scanln(&overwriteMPM)
-			if overwriteMPM == "n" || overwriteMPM == "N" {
-				fmt.Println("Skipping download.")
-				if runtime.GOOS == "windows" || runtime.GOOS == "darwin" {
-					unzipPath := filepath.Join(mpmDownloadPath, "mpm-contents")
+		for {
+			if err == nil {
+				fmt.Print("MPM already exists in this directory. Would you like to overwrite it? ")
+				fmt.Print(red("This will also overwrite the directory 'mpm-contents' if it already exists. (y/n)\n> "))
+				var overwriteMPM string
+				fmt.Scanln(&overwriteMPM)
+				if overwriteMPM == "n" || overwriteMPM == "N" {
+					fmt.Println("Skipping download.")
+					mpmDownloadNeeded = false
+					if runtime.GOOS == "windows" || runtime.GOOS == "darwin" {
+						unzipPath := filepath.Join(mpmDownloadPath, "mpm-contents")
 
-					// Skip download if you want to use your existing MPM, but it is not extracted.
-					if _, err := os.Stat(unzipPath); os.IsNotExist(err) {
-						mpmDownloadNeeded = false
+						// Skip download if you want to use your existing MPM, but it is not extracted.
+						if _, err := os.Stat(unzipPath); os.IsNotExist(err) {
+							break
+						} else {
+							fmt.Println("Skipping extraction.")
+							mpmExtractNeeded = false
+							break
+						}
 					} else {
+						mpmExtractNeeded = false
 						break
 					}
-				} else {
+				}
+				if overwriteMPM == "y" || overwriteMPM == "Y" {
 					break
+				} else {
+					fmt.Println(red("Invalid choice. Please enter either 'y' or 'n'."))
+					continue
 				}
 			}
 		}
@@ -113,40 +136,42 @@ func main() {
 			fmt.Println("Beginning download of MPM. Please wait.")
 			err = downloadFile(mpmURL, fileName)
 			if err != nil {
-				fmt.Println("Failed to download MPM. ", err)
+				fmt.Println(red("Failed to download MPM. ", err))
 				continue // Go back to the beginning of the loop
 			}
 			fmt.Println("MPM downloaded successfully.")
 		}
 
 		// Unzip the file if using Windows or macOS.
-		if runtime.GOOS == "windows" || runtime.GOOS == "darwin" {
-			fmt.Println("Beginning extraction of MPM.")
-			unzipPath := filepath.Join(mpmDownloadPath, "mpm-contents")
+		if mpmExtractNeeded {
+			if runtime.GOOS == "windows" || runtime.GOOS == "darwin" {
+				fmt.Println("Beginning extraction of MPM.")
+				unzipPath := filepath.Join(mpmDownloadPath, "mpm-contents")
 
-			// Check if the "mpm-contents" directory already exists.
-			if _, err := os.Stat(unzipPath); err == nil {
+				// Check if the "mpm-contents" directory already exists.
+				if _, err := os.Stat(unzipPath); err == nil {
 
-				// Delete the existing "mpm-contents" directory if it's there.
-				err := os.RemoveAll(unzipPath)
+					// Delete the existing "mpm-contents" directory if it's there.
+					err := os.RemoveAll(unzipPath)
+					if err != nil {
+						fmt.Println(red("Failed to delete the existing 'mpm-contents' directory:", err))
+						continue // Go back to the beginning of the loop
+					}
+				}
+
+				err := os.MkdirAll(unzipPath, 0755)
 				if err != nil {
-					fmt.Println("Failed to delete the existing 'mpm-contents' directory:", err)
+					fmt.Println(red("Failed to create the directory:", err))
 					continue // Go back to the beginning of the loop
 				}
-			}
 
-			err := os.MkdirAll(unzipPath, 0755)
-			if err != nil {
-				fmt.Println("Failed to create the directory:", err)
-				continue // Go back to the beginning of the loop
+				err = unzipFile(fileName, unzipPath)
+				if err != nil {
+					fmt.Println(red("Failed to extract MPM:", err))
+					continue // Go back to the beginning of the loop
+				}
+				fmt.Println("MPM extracted successfully.")
 			}
-
-			err = unzipFile(fileName, unzipPath)
-			if err != nil {
-				fmt.Println("Failed to extract MPM:", err)
-				continue // Go back to the beginning of the loop
-			}
-			fmt.Println("MPM extracted successfully.")
 		}
 		break // Exit the loop
 	}
@@ -161,6 +186,7 @@ func main() {
 
 	for {
 		fmt.Printf("Enter which release you would like to install. Press Enter to select %s: ", defaultRelease)
+		fmt.Print("\n> ")
 		release, _ = reader.ReadString('\n')
 		release = strings.TrimSpace(release)
 		if release == "" {
@@ -182,13 +208,12 @@ func main() {
 			break
 		}
 
-		fmt.Println("Invalid release. Enter a release between R2017b-R2023b.")
+		fmt.Println(red("Invalid release. Enter a release between R2017b-R2023b."))
 	}
-	fmt.Println("release selected outside of loop:", release)
 
 	//Product selection.
-	fmt.Println("Enter the products you would like to install. Use the same syntax as MPM to specify products. Press Enter to install all products.")
-
+	fmt.Print("Enter the products you would like to install. Use the same syntax as MPM to specify products. " +
+		"Press Enter to install all products.\n> ")
 	productsInput, _ := reader.ReadString('\n')
 	productsInput = strings.TrimSpace(productsInput)
 
@@ -202,12 +227,7 @@ func main() {
 
 	fmt.Println("Products to install:", products)
 
-	fmt.Println("Enter the full path where you would like to install these products. Press Enter to install to default path.")
-
-	installPath, _ := reader.ReadString('\n')
-	installPath = strings.TrimSpace(installPath)
-
-	// Set the path based on your OS.
+	// Set the default installation path based on your OS.
 	if operatingSystem == "maci64" {
 		defaultPath = "/Applications/MATLAB_" + release
 	}
@@ -217,11 +237,57 @@ func main() {
 	if operatingSystem == "glnxa64" {
 		defaultPath = "/usr/local/MATLAB/" + release
 	}
+
+	fmt.Print("Enter the full path where you would like to install these products. "+
+		"Press Enter to install to default path: \"", defaultPath, "\"\n> ")
+
+	installPath, _ := reader.ReadString('\n')
+	installPath = strings.TrimSpace(installPath)
+
 	if installPath == "" {
 		installPath = defaultPath
 	}
 
+	// Add some code to check the following:
+	// - If you have permissions to read/write there
+
 	fmt.Println("Installation path:", installPath)
+
+	// Option license file selection.
+	fmt.Print("If you have a license file you'd like to include in your installation, " +
+		"please provide the full path to the existing license file.\n> ")
+
+	licensePath, _ := reader.ReadString('\n')
+	licensePath = strings.TrimSpace(licensePath)
+
+	// Add some code...
+	// - To check if the file actually exists
+	// - That sets licenseUse = true
+	// - Does not accept ""
+
+	var mpmFullPath string
+	fmt.Println(licensePath)
+
+	if runtime.GOOS == "darwin" {
+		mpmFullPath = mpmDownloadPath + "//mpm-contents//bin//maci64//mpm"
+	}
+	if runtime.GOOS == "windows" {
+		mpmFullPath = mpmDownloadPath + "\\mpm-contents\\bin\\win64\\mpm.exe"
+	}
+	if runtime.GOOS == "linux" {
+		mpmFullPath = mpmDownloadPath + "/mpm"
+	}
+	fmt.Println(mpmFullPath)
+
+	// Yes, this is broken. Will come back to it tomorrow.
+	cmd := exec.Command(mpmFullPath, "install --release=", release, " --destination=", installPath, " --products ")
+	cmd.Stdout = os.Stdout
+	cmd.Stderr = os.Stderr
+	err := cmd.Run()
+	if err != nil {
+		// Handle any errors that occurred during execution
+		fmt.Println("Error executing mpm:", err)
+	}
 
 	// Next steps:
 	// - May need to chmod mpm on Linux. Should test this soon.
